Update modify.py
Removed pairwise distance calculation. Only evolutionary distance to the original is computed. Also, the ratio is stored instead of the actual number of differences. The -l argument to the script is now a percentage of the original instead of a whole number. E.g., "-l 0.3" will modify 30 percent of the letters.